# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.2.0
  description: MoveEasy API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication
  - name: moves
    description: Move cases
  - name: providers
    description: Providers
  - name: admin
    description: Admin operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterBody"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Email already in use
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Log in
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Log out
      responses:
        "204":
          description: Logged out

  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /providers:
    get:
      operationId: listProviders
      tags: [providers]
      summary: List all active providers
      responses:
        "200":
          description: List of providers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Provider"

  /moves:
    get:
      operationId: listMoves
      tags: [moves]
      summary: List moves for current user
      responses:
        "200":
          description: List of moves
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Move"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createMove
      tags: [moves]
      summary: Create a new move
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMoveBody"
      responses:
        "201":
          description: Move created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Move"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /moves/{id}:
    get:
      operationId: getMove
      tags: [moves]
      summary: Get move details with provider statuses
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Move detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MoveDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /moves/{id}/providers/{providerId}:
    patch:
      operationId: updateMoveProvider
      tags: [moves]
      summary: Update provider status for a move
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: providerId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMoveProviderBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MoveProviderStatus"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /admin/stats:
    get:
      operationId: adminGetStats
      tags: [admin]
      summary: Get admin overview stats
      responses:
        "200":
          description: Stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminStats"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /admin/cases:
    get:
      operationId: adminListCases
      tags: [admin]
      summary: List all move cases (admin)
      parameters:
        - name: search
          in: query
          required: false
          schema:
            type: string
        - name: status
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of all cases
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminCase"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /admin/cases/{id}:
    get:
      operationId: adminGetCase
      tags: [admin]
      summary: Get full case detail (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Case detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminCaseDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /admin/providers:
    get:
      operationId: adminListProviders
      tags: [admin]
      summary: List all providers (admin)
      responses:
        "200":
          description: All providers including inactive
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Provider"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /admin/providers/{id}:
    patch:
      operationId: adminUpdateProvider
      tags: [admin]
      summary: Update a provider (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminUpdateProviderBody"
      responses:
        "200":
          description: Updated provider
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Provider"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /admin/settings:
    get:
      operationId: adminGetSettings
      tags: [admin]
      summary: Get admin settings
      responses:
        "200":
          description: Settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSettings"
    put:
      operationId: adminUpdateSettings
      tags: [admin]
      summary: Update admin settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminUpdateSettingsBody"
      responses:
        "200":
          description: Updated settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSettings"

  /admin/users/{id}/test:
    patch:
      operationId: adminToggleTestUser
      tags: [admin]
      summary: Toggle test user flag
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminToggleTestUserBody"
      responses:
        "200":
          description: Updated user test flag
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminToggleTestUserResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    RegisterBody:
      type: object
      properties:
        fullName:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
      required:
        - fullName
        - email
        - password

    LoginBody:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required:
        - email
        - password

    AuthUser:
      type: object
      properties:
        id:
          type: integer
        fullName:
          type: string
        email:
          type: string
        isAdmin:
          type: boolean
      required:
        - id
        - fullName
        - email
        - isAdmin

    Provider:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        category:
          type: string
        description:
          type: string
          nullable: true
        isAffiliate:
          type: boolean
        affiliateUrl:
          type: string
          nullable: true
        isActive:
          type: boolean
        logoUrl:
          type: string
          nullable: true
      required:
        - id
        - name
        - category
        - isAffiliate
        - isActive

    CreateMoveBody:
      type: object
      properties:
        oldAddressLine1:
          type: string
        oldAddressLine2:
          type: string
          nullable: true
        oldCity:
          type: string
        oldPostcode:
          type: string
        newAddressLine1:
          type: string
        newAddressLine2:
          type: string
          nullable: true
        newCity:
          type: string
        newPostcode:
          type: string
        moveDate:
          type: string
        providerIds:
          type: array
          items:
            type: integer
      required:
        - oldAddressLine1
        - oldCity
        - oldPostcode
        - newAddressLine1
        - newCity
        - newPostcode
        - moveDate
        - providerIds

    Move:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        oldAddressLine1:
          type: string
        oldAddressLine2:
          type: string
          nullable: true
        oldCity:
          type: string
        oldPostcode:
          type: string
        newAddressLine1:
          type: string
        newAddressLine2:
          type: string
          nullable: true
        newCity:
          type: string
        newPostcode:
          type: string
        moveDate:
          type: string
        status:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - oldAddressLine1
        - oldCity
        - oldPostcode
        - newAddressLine1
        - newCity
        - newPostcode
        - moveDate
        - status
        - createdAt
        - updatedAt

    MoveProviderStatus:
      type: object
      properties:
        id:
          type: integer
        moveId:
          type: integer
        providerId:
          type: integer
        providerName:
          type: string
        providerCategory:
          type: string
        isAffiliate:
          type: boolean
        affiliateUrl:
          type: string
          nullable: true
        status:
          type: string
        notes:
          type: string
          nullable: true
        updatedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - moveId
        - providerId
        - providerName
        - providerCategory
        - isAffiliate
        - status
        - createdAt
        - updatedAt

    MoveDetail:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        oldAddressLine1:
          type: string
        oldAddressLine2:
          type: string
          nullable: true
        oldCity:
          type: string
        oldPostcode:
          type: string
        newAddressLine1:
          type: string
        newAddressLine2:
          type: string
          nullable: true
        newCity:
          type: string
        newPostcode:
          type: string
        moveDate:
          type: string
        status:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        providers:
          type: array
          items:
            $ref: "#/components/schemas/MoveProviderStatus"
      required:
        - id
        - userId
        - oldAddressLine1
        - oldCity
        - oldPostcode
        - newAddressLine1
        - newCity
        - newPostcode
        - moveDate
        - status
        - createdAt
        - updatedAt
        - providers

    UpdateMoveProviderBody:
      type: object
      properties:
        status:
          type: string
          enum: [pending, submitted, completed, failed, action_required]
        notes:
          type: string
          nullable: true
      required:
        - status

    AdminStats:
      type: object
      properties:
        totalMoves:
          type: integer
        activeMoves:
          type: integer
        completedMoves:
          type: integer
        totalUsers:
          type: integer
        totalProviders:
          type: integer
        pendingSubmissions:
          type: integer
        actionRequired:
          type: integer
      required:
        - totalMoves
        - activeMoves
        - completedMoves
        - totalUsers
        - totalProviders
        - pendingSubmissions
        - actionRequired

    AdminCase:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        userEmail:
          type: string
        userFullName:
          type: string
        oldAddressLine1:
          type: string
        oldCity:
          type: string
        oldPostcode:
          type: string
        newAddressLine1:
          type: string
        newCity:
          type: string
        newPostcode:
          type: string
        moveDate:
          type: string
        status:
          type: string
        providerCount:
          type: integer
        pendingCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - userEmail
        - userFullName
        - oldAddressLine1
        - oldCity
        - oldPostcode
        - newAddressLine1
        - newCity
        - newPostcode
        - moveDate
        - status
        - providerCount
        - pendingCount
        - createdAt
        - updatedAt

    AdminCaseDetail:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        userEmail:
          type: string
        userFullName:
          type: string
        isTestUser:
          type: boolean
        oldAddressLine1:
          type: string
        oldAddressLine2:
          type: string
          nullable: true
        oldCity:
          type: string
        oldPostcode:
          type: string
        newAddressLine1:
          type: string
        newAddressLine2:
          type: string
          nullable: true
        newCity:
          type: string
        newPostcode:
          type: string
        moveDate:
          type: string
        status:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        providers:
          type: array
          items:
            $ref: "#/components/schemas/MoveProviderStatus"
      required:
        - id
        - userId
        - userEmail
        - userFullName
        - oldAddressLine1
        - oldCity
        - oldPostcode
        - newAddressLine1
        - newCity
        - newPostcode
        - moveDate
        - status
        - createdAt
        - updatedAt
        - providers

    AdminUpdateProviderBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
          nullable: true
        isAffiliate:
          type: boolean
        affiliateUrl:
          type: string
          nullable: true
        isActive:
          type: boolean

    AdminSettings:
      type: object
      properties:
        testPreviewEmail:
          type: string
      required:
        - testPreviewEmail

    AdminUpdateSettingsBody:
      type: object
      properties:
        testPreviewEmail:
          type: string
      required:
        - testPreviewEmail

    AdminToggleTestUserBody:
      type: object
      properties:
        isTestUser:
          type: boolean
      required:
        - isTestUser

    AdminToggleTestUserResponse:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        isTestUser:
          type: boolean
      required:
        - id
        - email
        - isTestUser