# lib/api-zod/src/generated/types/adminCaseDetail.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MoveEasy API specification
 * OpenAPI spec version: 0.2.0
 */
import type { MoveProviderStatus } from "./moveProviderStatus";

export interface AdminCaseDetail {
  id: number;
  userId: number;
  userEmail: string;
  userFullName: string;
  isTestUser?: boolean;
  oldAddressLine1: string;
  oldAddressLine2?: string | null;
  oldCity: string;
  oldPostcode: string;
  newAddressLine1: string;
  newAddressLine2?: string | null;
  newCity: string;
  newPostcode: string;
  moveDate: string;
  status: string;
  createdAt: Date;
  updatedAt: Date;
  providers: MoveProviderStatus[];
}